requirements.txt
flask
qrcode
markupsafe
requests
openpyxl
python-docx
pyinstaller
python-dotenv